controller, Service, entity 수정 config 관련 확인 필요

+++ src/main/java/com/projet/clubpage/entity/File.java
@@ -17,7 +17,7 @@ public class File {
 
     @Id
     @GeneratedValue
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false)
     private String origFilename;
@@ -31,7 +31,7 @@ public class File {
 
 
     @Builder
-    public File(Long id, String origFilename, String filename, String filePath) {
+    public File(Integer id, String origFilename, String filename, String filePath) {
         this.id = id;
         this.origFilename = origFilename;
         this.filename = filename;
